+++ core-sdk/src/main/java/com/iov42/solutions/core/sdk/HttpHostWrapper.java
@@ -1,6 +1,6 @@
 package com.iov42.solutions.core.sdk;
 
-import com.iov42.solutions.core.sdk.http.HttpBackend;
+import com.iov42.solutions.core.sdk.http.spi.HttpBackend;
 import com.iov42.solutions.core.sdk.http.HttpBackendException;
 import com.iov42.solutions.core.sdk.http.HttpBackendRequest;
 import com.iov42.solutions.core.sdk.http.HttpBackendResponse;
@@ -10,6 +10,8 @@ import com.iov42.solutions.core.sdk.utils.serialization.*;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
+import java.time.Duration;
+import java.time.Instant;
 import java.util.List;
 import java.util.Map;
 import java.util.concurrent.CompletableFuture;
@@ -19,8 +21,7 @@ class HttpHostWrapper implements HttpBackend {
 
     private static final Logger log = LoggerFactory.getLogger(HttpHostWrapper.class);
 
-    private static final int MAX_REDIRECT_COUNT = 5;
-
+    private static final Duration RETRY_TIMEOUT = Duration.ofSeconds(40);
     private final HttpBackend httpBackend;
     private final String platformHost;
 
@@ -31,7 +32,7 @@ class HttpHostWrapper implements HttpBackend {
 
     public <T> T executeGet(String url, Map<String, List<String>> headers, Class<T> responseClass) {
         try {
-            return execute(1, new HttpBackendRequest(HttpBackendRequest.Method.GET, url, headers, null), responseClass).join();
+            return execute(Instant.now(), new HttpBackendRequest(HttpBackendRequest.Method.GET, url, headers, null), responseClass).join();
         } catch (CompletionException ex) {
             if (ex.getCause() instanceof RuntimeException) {
                 throw (RuntimeException) ex.getCause();
@@ -41,7 +42,7 @@ class HttpHostWrapper implements HttpBackend {
     }
 
     public <T> CompletableFuture<T> executePut(String url, byte[] body, Map<String, List<String>> headers, Class<T> responseClass) {
-        return execute(1, new HttpBackendRequest(HttpBackendRequest.Method.PUT, url, headers, body), responseClass);
+        return execute(Instant.now(), new HttpBackendRequest(HttpBackendRequest.Method.PUT, url, headers, body), responseClass);
     }
 
     @Override
@@ -55,14 +56,17 @@ class HttpHostWrapper implements HttpBackend {
         return httpBackend.execute(requestClone).thenApply(r -> r != null ? r.mutate(request) : null);
     }
 
-    private <T> CompletableFuture<T> execute(int redirectCount, HttpBackendRequest request, Class<T> responseClass) throws HttpBackendException {
-        return execute(request).thenCompose(r -> handleResponse(redirectCount, r, responseClass));
+    private <T> CompletableFuture<T> execute(Instant start, HttpBackendRequest request, Class<T> responseClass) throws HttpBackendException {
+        return execute(request).thenCompose(r -> handleResponse(start, r, responseClass));
     }
 
-    private <T> CompletableFuture<T> handleResponse(int attempt, HttpBackendResponse response, Class<T> responseClass) {
-        if (attempt < MAX_REDIRECT_COUNT) {
-            int statusCode = response.statusCode();
-            if (statusCode >= 300 && statusCode <= 399) {
+    private <T> CompletableFuture<T> handleResponse(Instant start, HttpBackendResponse response, Class<T> responseClass) {
+        int statusCode = response.statusCode();
+        if (statusCode >= 300 && statusCode <= 399) {
+            // we are asked to redirect
+            Duration waited = Duration.between(start, Instant.now());
+            if (waited.compareTo(RETRY_TIMEOUT) < 0) {
+
                 String location = getValue(response.headers().get("Location"));
                 String retryAfter = getValue(response.headers().get("Retry-After"));
                 if (retryAfter != null) {
@@ -75,7 +79,10 @@ class HttpHostWrapper implements HttpBackend {
                 }
                 // build redirect request and execute
                 HttpBackendRequest redirectRequest = response.getRequest().mutate(location);
-                return execute(attempt + 1, redirectRequest, responseClass);
+                log.trace("Retrying request {}", response.getRequest().getRequestUrl());
+                return execute(start, redirectRequest, responseClass);
+            } else {
+                throw new PlatformRetryExceededException();
             }
         }
         return CompletableFuture.completedFuture(convertResponse(response, responseClass));
@@ -95,6 +102,9 @@ class HttpHostWrapper implements HttpBackend {
             }
             ErrorResponse errorResponse = convertResponse(response, ErrorResponse.class);
             throw new PlatformErrorException(statusCode, errorResponse);
+        }
+        if (log.isDebugEnabled()) {
+            log.debug("Http Response. Url: {}, Status: {}, Body: {}", response.getRequest().getRequestUrl(), response.statusCode(), response.body());
         }
         return JsonUtils.fromJson(response.body(), clazz);
     }