+++ core-sdk/src/main/java/com/iov42/solutions/core/sdk/PlatformRetryExceededException.java
@@ -1,0 +1,4 @@
+package com.iov42.solutions.core.sdk;
+
+public class PlatformRetryExceededException extends RuntimeException {
+}