+++ core-sdk/src/test/java/com/iov42/solutions/core/sdk/HttpHostWrapperTest.java
@@ -1,6 +1,6 @@
 package com.iov42.solutions.core.sdk;
 
-import com.iov42.solutions.core.sdk.http.HttpBackend;
+import com.iov42.solutions.core.sdk.http.spi.HttpBackend;
 import com.iov42.solutions.core.sdk.http.HttpBackendRequest;
 import com.iov42.solutions.core.sdk.http.HttpBackendResponse;
 import com.iov42.solutions.core.sdk.model.PlatformError;